requirements.txt
paho-mqtt>=1.6.1,<3.0
PyYAML>=6.0
